+++ changelog.md
@@ -3,6 +3,8 @@
 - <c-dddddd>Added button and keybind to move selected objects to your editor position</c>
 - <c-dddddd>Added button and keybind to move your editor position to selected objects</c>
 - <c-dddddd>Make letting go of shift before mouse when swiping stopping swipe an option</c>
+- <c-dddddd>Hackily make quick volume controls work with my scroll rewrite</c>
+- <c-dddddd>Block scroll when paused</c>
 
 # 1.0.0-beta.23
 - <c-dddddd>Make object search items load while in the editor rather than on level load. Load rate can be customized in settings</c>